+++ test/cases/option-output-public-path-custom/webpack.config.js
@@ -1,0 +1,58 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+    publicPath: '/sub-path/',
+  },
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      entry: {
+        index: {
+          import: './src/index.html',
+          data: {
+            title: 'Home',
+            name: 'World',
+          },
+        },
+      },
+
+      js: {
+        // output filename of extracted JS
+        filename: 'assets/js/[name].[contenthash:8].js',
+      },
+
+      css: {
+        // output filename of extracted CSS
+        filename: 'assets/css/[name].[contenthash:8].css',
+      },
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.html$/,
+        loader: HtmlBundlerPlugin.loader,
+      },
+
+      {
+        test: /.css$/,
+        use: ['css-loader'],
+      },
+
+      {
+        test: /.(png|jpe?g|ico|svg)/,
+        type: 'asset/resource',
+        generator: {
+          //filename: 'assets/img/[name].[hash:8][ext]',
+          filename: 'assets/img/[name].[hash:8][ext]',
+        },
+      },
+    ],
+  },
+};