+++ test/config.js
@@ -4,7 +4,6 @@ const PATHS = {
   base: __dirname,
   testSource: path.join(__dirname, 'cases'),
   // relative path in the test directory to web root dir name, same as by a web server (e.g. nginx)
-  //webRoot: '/public/',
   webRoot: '/dist/',
   // relative path in the test directory to expected files for test
   expected: '/expected/',